HW 3: refactor, extract errorHandling into separate middleware
