+++ src/middleware/errorHandling/errorHandling.js
@@ -1,0 +1,3 @@
+export const consoleErrorhandler = err => {
+    console.error(err);
+};
